Create DTO (Create EmployeeServiceImpl & interface EmployeeService)
